+++ src/collective/documentgenerator/helper/archetypes.py
@@ -12,31 +12,28 @@ class ATDocumentGenerationHelperView(DocumentGenerationHelperView):
     Archetypes implementation of document generation helper methods.
     """
 
-    def display(self, field_name, context=None, no_value=''):
-        if context is None:
-            context = self.real_context
-
-        if self.check_permission(field_name, context):
-            field_renderer = self.get_AT_field_renderer(field_name, context)
+    def display(self, field_name, no_value=''):
+        if self.check_permission(field_name):
+            field_renderer = self.get_AT_field_renderer(field_name)
             display_value = field_renderer.render(no_value=no_value)
         else:
             display_value = u''
 
         return display_value
 
-    def check_permission(self, field_name, context):
-        return bool(context.getField(field_name).checkPermission('r', context))
+    def check_permission(self, field_name):
+        return bool(self.real_context.getField(field_name).checkPermission('r', self.real_context))
 
-    def get_AT_field_renderer(self, field_name, context):
-        field = context.getField(field_name)
+    def get_AT_field_renderer(self, field_name):
+        field = self.real_context.getField(field_name)
         widget = field.widget
-        renderer = getMultiAdapter((field, widget, context), IFieldRendererForDocument)
+        renderer = getMultiAdapter((field, widget, self.real_context), IFieldRendererForDocument)
 
         return renderer
 
-    def display_date(self, field_name, context=None, long_format=None, time_only=None, custom_format=None):
-        date_field = self.context.getField(field_name)
-        date = date_field.get(self.context)
+    def display_date(self, field_name, long_format=None, time_only=None, custom_format=None):
+        date_field = self.real_context.getField(field_name)
+        date = date_field.get(self.real_context)
         if not custom_format:
             # use toLocalizedTime
             formatted_date = self.plone.toLocalizedTime(date, long_format, time_only)
@@ -45,29 +42,23 @@ class ATDocumentGenerationHelperView(DocumentGenerationHelperView):
 
         return formatted_date
 
-    def display_voc(self, field_name, context=None, separator=', '):
-        if context is None:
-            context = self.real_context
+    def display_voc(self, field_name, separator=', '):
+        display_value = self.real_context.restrictedTraverse('@@at_utils').translate
 
-        display_value = context.restrictedTraverse('@@at_utils').translate
-
-        field = context.getField(field_name)
-        voc = field.Vocabulary(context)
-        raw_values = field.get(context)
+        field = self.real_context.getField(field_name)
+        voc = field.Vocabulary(self.real_context)
+        raw_values = field.get(self.real_context)
         values = [display_value(voc, val) for val in raw_values]
         display = separator.join(values)
 
         return display
 
-    def display_text(self, field_name, context=None):
+    def display_text(self, field_name):
         if not self.appy_renderer:
             return ''
 
-        if context is None:
-            context = self.real_context
-
-        html_field = self.context.getField(field_name)
-        html_text = html_field.get(context)
+        html_field = self.real_context.getField(field_name)
+        html_text = html_field.get(self.real_context)
         display = self.appy_renderer.renderXhtml(html_text)
         return display